Replace system call with in-house bash
C++ "system()" function uses the OS's non-interactive shell to execute the command.
In a lot of places, this is "sh", and not bash (even though the .sh scripts contain the
"#!/bin/bash" shebang).

The problem with this is that sh doesn't support a lot of bash syntax. Case in point,
if you run libbashbullet as is on a Debian 9.0 system, no script will run because of
"Syntax error: redirection unexpected" error.

Replace this by forcefully executing "bin/bash" using execl, and scripts will now work.

This could also become configurable in the config.json